+++ src/api/users/personal_access_tokens.rs
@@ -16,8 +16,9 @@ mod create_for_user;
 pub use self::create::CreatePersonalAccessToken;
 pub use self::create::CreatePersonalAccessTokenBuilder;
 pub use self::create::CreatePersonalAccessTokenBuilderError;
-pub use self::create::PersonalAccessTokenScope;
+pub use self::create::PersonalAccessTokenCreateScope;
 
 pub use self::create_for_user::CreatePersonalAccessTokenForUser;
 pub use self::create_for_user::CreatePersonalAccessTokenForUserBuilder;
 pub use self::create_for_user::CreatePersonalAccessTokenForUserBuilderError;
+pub use self::create_for_user::PersonalAccessTokenScope;