+++ src/api/users/personal_access_tokens/create_for_user.rs
@@ -10,7 +10,74 @@ use chrono::NaiveDate;
 use derive_builder::Builder;
 
 use crate::api::endpoint_prelude::*;
-use crate::api::users::personal_access_tokens::PersonalAccessTokenScope;
+use crate::api::ParamValue;
+
+/// Scopes for personal access tokens.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[non_exhaustive]
+pub enum PersonalAccessTokenScope {
+    /// Access the API and perform git reads and writes.
+    Api,
+    /// Access to read the user information.
+    ReadUser,
+    /// Access read-only API endpoints.
+    ReadApi,
+    /// Read access to repositories.
+    ReadRepository,
+    /// Write access to repositories.
+    WriteRepository,
+    /// Read access to Docker registries.
+    ReadRegistry,
+    /// Write access to Docker registries.
+    WriteRegistry,
+    /// Permission to `sudo` as other users (administrator only).
+    Sudo,
+    /// Permission to access administrator API actions.
+    AdminMode,
+    /// Permission to create instance runners.
+    CreateRunner,
+    /// Access to AI features (GitLab Duo for JetBrains).
+    AiFeatures,
+    /// Access to perform Kubernetes API calls.
+    K8sProxy,
+    /// Access to the Service Ping payload.
+    ReadServicePing,
+}
+
+impl PersonalAccessTokenScope {
+    /// The scope as a query parameter.
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::Api => "api",
+            Self::ReadUser => "read_user",
+            Self::ReadApi => "read_api",
+            Self::ReadRepository => "read_repository",
+            Self::WriteRepository => "write_repository",
+            Self::ReadRegistry => "read_registry",
+            Self::WriteRegistry => "write_registry",
+            Self::Sudo => "sudo",
+            Self::AdminMode => "admin_mode",
+            Self::CreateRunner => "create_runner",
+            Self::AiFeatures => "ai_features",
+            Self::K8sProxy => "k8s_proxy",
+            Self::ReadServicePing => "read_service_ping",
+        }
+    }
+
+    /// Transform into a `create` scope if possible.
+    pub fn as_create_scope(self) -> Option<super::PersonalAccessTokenCreateScope> {
+        match self {
+            Self::K8sProxy => Some(super::PersonalAccessTokenCreateScope::K8sProxy),
+            _ => None,
+        }
+    }
+}
+
+impl ParamValue<'static> for PersonalAccessTokenScope {
+    fn as_value(&self) -> Cow<'static, str> {
+        self.as_str().into()
+    }
+}
 
 /// Create a new personal access token for a user.
 #[derive(Debug, Builder, Clone)]
@@ -87,6 +154,35 @@ mod tests {
     };
     use crate::api::{self, Query};
     use crate::test::client::{ExpectedUrl, SingleTestClient};
+
+    #[test]
+    fn personal_access_token_scope_as_str() {
+        let items = &[
+            (PersonalAccessTokenScope::Api, "api"),
+            (PersonalAccessTokenScope::ReadUser, "read_user"),
+            (PersonalAccessTokenScope::ReadApi, "read_api"),
+            (PersonalAccessTokenScope::ReadRepository, "read_repository"),
+            (
+                PersonalAccessTokenScope::WriteRepository,
+                "write_repository",
+            ),
+            (PersonalAccessTokenScope::ReadRegistry, "read_registry"),
+            (PersonalAccessTokenScope::WriteRegistry, "write_registry"),
+            (PersonalAccessTokenScope::Sudo, "sudo"),
+            (PersonalAccessTokenScope::AdminMode, "admin_mode"),
+            (PersonalAccessTokenScope::CreateRunner, "create_runner"),
+            (PersonalAccessTokenScope::AiFeatures, "ai_features"),
+            (PersonalAccessTokenScope::K8sProxy, "k8s_proxy"),
+            (
+                PersonalAccessTokenScope::ReadServicePing,
+                "read_service_ping",
+            ),
+        ];
+
+        for (i, s) in items {
+            assert_eq!(i.as_str(), *s);
+        }
+    }
 
     #[test]
     fn user_name_and_scopes_are_necessary() {